Update apollo client createIsoMorphic link and generated graphql schema
